Cause cache-clearing to run on the server
